Move the application logic into its own function
This allows us to pass configuration values when starting the application
